fixup! feat: add vip for Chaos labs recommendations for staked eth pool
