+++ multisig/proposals/ethereum/vip-063/index.ts
@@ -5,7 +5,7 @@ import {
   LIQUID_STAKED_ETH_COMPTROLLER,
   NINETY_PERCENT_CF,
   vwETH,
-} from "../../../../vips/vip-374/bscmainnet";
+} from "../../../../vips/vip-377/bscmainnet";
 
 export const vip063 = () => {
   return makeProposal([
